Made trim work with really large files. (#470)

+++ src/libasf_raster/trim.c
@@ -115,7 +115,7 @@ int trim(char *infile, char *outfile,
     offset=0;
     
     for (;y<lastReadY;y++) {
-      int inputY=y+startY,
+      long long int inputY=y+startY,
         inputX=firstReadX+startX,
         outputX=firstReadX;
